Move from_iter_values to GenericByteArray
